integrated remote script into function, but not main game code yet
