+++ setup.sh
@@ -2,7 +2,7 @@
 
 #list name of packages
 echo "installing base packages"
-conda create -y -n hitw python=3.6 numpy pandas joblib pyaudio smbus2 paho-mqtt scikit-learn opencv speechrecognition boto3 -c conda-forge
+conda create -y -n hitw python=3.6 numpy pandas joblib pyaudio smbus2 paho-mqtt scikit-learn opencv speechrecognition boto3 paramiko -c conda-forge
 conda activate hitw 
 echo "finished"
